Refactor: Extract login information validation method
