+++ pages/helpers/users/index.ts
@@ -1,13 +1,15 @@
-import getUserByEmail from "./getUserByEmail";
-import { validatePassword } from "./validatePassword";
-import validateCookie from "./validateCookie";
-import getUserByEmailWithoutPassword from "./getUserByEmailWithoutPassword";
-import getAllProjects from "./getAllProjects";
+import getUserByEmail from './getUserByEmail'
+import { validatePassword } from './validatePassword'
+import validateCookie from './validateCookie'
+import getUserByEmailWithoutPassword from './getUserByEmailWithoutPassword'
+import getAllProjects from './getAllProjects'
+import validateLoginInformation from './validateLoginInformation'
 
 export {
-    getUserByEmail,
-    validatePassword,
-    validateCookie,
-    getUserByEmailWithoutPassword,
-    getAllProjects,
-};
+  getUserByEmail,
+  validatePassword,
+  validateCookie,
+  getUserByEmailWithoutPassword,
+  getAllProjects,
+  validateLoginInformation,
+}
